requirements.txt
streamlit
insightface
onnxruntime
scikit-learn
opencv-python
pillow
numpy
tqdm
